passage en float de certaines coord pour vitesses en decimal, mais segfault

+++ Srcs/hooks_changes.c
@@ -6,7 +6,7 @@
 /*   By: tgellon <tgellon@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/16 10:41:04 by rrebois           #+#    #+#             */
-/*   Updated: 2023/09/18 14:18:23 by tgellon          ###   ########lyon.fr   */
+/*   Updated: 2023/09/19 10:34:58 by tgellon          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,11 +75,11 @@ void	move_left(t_data *data)
 				data->player.dir.y * cos(-M_PI_2);
 	new_pos.x = new_dir.x * move_speed;
 	new_pos.y = new_dir.y * move_speed;
-	if (data->map.map[data->player.pos.y / SQUARE_SIZE][(data->player.pos.x + \
+	if (data->map.map[(int)data->player.pos.y / SQUARE_SIZE][(int)(data->player.pos.x + \
 		new_pos.x) / SQUARE_SIZE] != '1')
 		data->player.pos.x += new_pos.x;
-	if (data->map.map[(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
-		[data->player.pos.x / SQUARE_SIZE] != '1')
+	if (data->map.map[(int)(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
+		[(int)data->player.pos.x / SQUARE_SIZE] != '1')
 		data->player.pos.y += new_pos.y;
 }
 
@@ -96,11 +96,11 @@ void	move_right(t_data *data)
 				data->player.dir.y * cos(-M_PI_2);
 	new_pos.x = new_dir.x * move_speed;
 	new_pos.y = new_dir.y * move_speed;
-	if (data->map.map[data->player.pos.y / SQUARE_SIZE][(data->player.pos.x + \
+	if (data->map.map[(int)data->player.pos.y / SQUARE_SIZE][(int)(data->player.pos.x + \
 		new_pos.x) / SQUARE_SIZE] != '1')
 		data->player.pos.x += new_pos.x;
-	if (data->map.map[(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
-		[data->player.pos.x / SQUARE_SIZE] != '1')
+	if (data->map.map[(int)(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
+		[(int)data->player.pos.x / SQUARE_SIZE] != '1')
 		data->player.pos.y += new_pos.y;
 }
 
@@ -112,12 +112,18 @@ void	move_forward(t_data *data)
 	move_speed = MOVE_SPEED;
 	new_pos.x = data->player.dir.x * move_speed;
 	new_pos.y = data->player.dir.y * move_speed;
-	if (data->map.map[data->player.pos.y / SQUARE_SIZE][(data->player.pos.x + \
+// printf("new.x :%f\n", data->player.pos.x);
+// printf("new.y :%f\n", data->player.pos.y);
+	if (data->map.map[(int)data->player.pos.y / SQUARE_SIZE][(int)(data->player.pos.x + \
 		new_pos.x) / SQUARE_SIZE] != '1')
 		data->player.pos.x += new_pos.x;
-	if (data->map.map[(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
-		[data->player.pos.x / SQUARE_SIZE] != '1')
+	if (data->map.map[(int)(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
+		[(int)data->player.pos.x / SQUARE_SIZE] != '1')
 		data->player.pos.y += new_pos.y;
+// printf("dir.x :%f\n", data->player.dir.x);
+// printf("dir.y :%f\n", data->player.dir.y);
+// printf("new.x :%f\n", data->player.pos.x);
+// printf("new.y :%f\n\n", data->player.pos.y);
 }
 
 void	move_backward(t_data *data)
@@ -128,10 +134,10 @@ void	move_backward(t_data *data)
 	move_speed = -MOVE_SPEED;
 	new_pos.x = data->player.dir.x * move_speed;
 	new_pos.y = data->player.dir.y * move_speed;
-	if (data->map.map[data->player.pos.y / SQUARE_SIZE][(data->player.pos.x + \
+	if (data->map.map[(int)data->player.pos.y / SQUARE_SIZE][(int)(data->player.pos.x + \
 		new_pos.x) / SQUARE_SIZE] != '1')
 		data->player.pos.x += new_pos.x;
-	if (data->map.map[(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
-		[data->player.pos.x / SQUARE_SIZE] != '1')
+	if (data->map.map[(int)(data->player.pos.y + new_pos.y) / SQUARE_SIZE] \
+		[(int)data->player.pos.x / SQUARE_SIZE] != '1')
 		data->player.pos.y += new_pos.y;
 }
